Complete redesign

+++ src/app/pages/profile/profile.component.ts
@@ -5,6 +5,7 @@ import { NewEventComponent } from '../../components/new-event/new-event.componen
 import { IEvent } from '../../shared/interfaces/event.interface';
 import { AuthService } from '../../shared/services/auth.service';
 import { EventService } from '../../shared/services/event.service';
+import { CopyEventComponent } from 'src/app/components/copy-event/copy-event.component';
 
 
 @Component({
@@ -19,8 +20,8 @@ export class ProfileComponent implements OnInit {
 
   constructor(
     public dialog: MatDialog,
-    private authService: AuthService, 
-    private eventService: EventService ) { }
+    private authService: AuthService,
+    private eventService: EventService) { }
 
   ngOnInit(): void {
     this.getChannelData()
@@ -37,8 +38,8 @@ export class ProfileComponent implements OnInit {
       events.forEach(event => {
         const eventData = event.data() as IEvent
         const id = event.id
-        const newEvent = {...eventData, id}
-        newEvent.startDate = new Date(newEvent.startDate['seconds']*1000)
+        const newEvent = { ...eventData, id }
+        newEvent.startDate = new Date(newEvent.startDate['seconds'] * 1000)
         this.events.push(newEvent)
       })
     })
@@ -46,16 +47,27 @@ export class ProfileComponent implements OnInit {
 
   openAddDialog() {
     const dialogRef = this.dialog.open(NewEventComponent, {
-      width: '450px',
+      width: '600px',
       data: { channel: this.channel, actionAdd: true }
     });
     dialogRef.afterClosed().subscribe(() => {
       this.loadEvents()
     });
   }
+
+  openCopyDialog(event: IEvent) {
+    const dialogRef = this.dialog.open(CopyEventComponent, {
+      width: '600px',
+      data: { event: event }
+    });
+    dialogRef.afterClosed().subscribe(() => {
+      this.loadEvents()
+    });
+  }
+
   openEditDialog(event: IEvent) {
     const dialogRef = this.dialog.open(NewEventComponent, {
-      width: '450px',
+      width: '600px',
       data: { channel: this.channel, actionAdd: false, event: event }
     });
     dialogRef.afterClosed().subscribe(() => {
@@ -64,9 +76,15 @@ export class ProfileComponent implements OnInit {
   }
 
   deleteEvent(event) {
-    this.eventService.deleteEvent(event).then(()=>{
-      this.loadEvents()
-    })
+    if (confirm("Are you sure?"))
+      this.eventService.deleteEvent(event).then(() => {
+        this.loadEvents()
+      })
+  }
+
+  edit() {
+    // To be implemented
+    console.log('to be implemented')
   }
 
   logout() {
